> [feat/fix] bid-service 타입 수정 및 폴백 API 구현

+++ services/bid-service/src/main/java/com/bankrupang/sanjijk/bid/infrastructure/scheduler/AuctionEndScheduler.java
@@ -59,7 +59,7 @@ public class AuctionEndScheduler {
                             auctionId,
                             hasBid,
                             hasBid ? highestBidderId : null,
-                            hasBid ? Long.parseLong(currentPrice) : null,
+                            hasBid ? Integer.parseInt(currentPrice) : null,
                             LocalDateTime.now()
                     );
                     try {
